Move TypeFactory into Domain
