+++ src/Domain/ParameterInfo.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Firehed\PhpLsp\Domain;
 
-use Firehed\PhpLsp\Utility\TypeFactory;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Param;
 use PhpParser\PrettyPrinter\Standard as PrettyPrinter;
